Localized all dependanies, resizable panes
Localized all dependanies (no remote includes) and added a HTML text editor based on the Bootstrap version of WYSIHTML5.

This is an interim measure I intend to replace it with a different, better supported editor (as fixing all the old resizing, auto scrolling and formatting bugs is a rabbit hole I want to avoid) but it's serviceable enough to use to build out the other functionality.