Markdown editor
* Replaced HTML editor with Markdown editor.

While WYSIWYG editing is more desirable, the editor it was using wasn't reliable enough. The Markdown editor is much more stable and was easy to integrate.

TinyMCE is the most likely candidate for a future full WYSIWYG editor, but probably it makes sense to leave in the option to use the Markdown editor too.

* Can now easily build mac-only test releases